+++ src/app/(home)/components/Nav/index.tsx
@@ -1,4 +1,4 @@
-import Tab from '@/components/common/Tab';
+import { Tab } from '@/components';
 import Image from 'next/image';
 import { useRouter } from 'next/navigation';
 import { useState } from 'react';
