+++ src/app/(home)/components/PlaceItem/index.tsx
