+++ src/app/(home)/components/index.ts
@@ -1,0 +1,2 @@
+export { default as Nav } from './Nav';
+export { default as PlaceItem } from './PlaceItem';
